requirements.txt
mcp[cli]>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pyyaml>=6.0
pytest>=7.0.0
